==> 0_Data-Preparation-Barcelona.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns

df = pd.read_csv("Datasets/Dataset_Originales/geographic_distribution_bcn.csv")
print(df.head())

df.columns

df.dtypes

# Count unique IDs in df
df['host_id'].nunique()

# See unique values in room_type column
df['room_type'].unique()

# Count number of each value in room_type column
df["room_type"].value_counts()

# View the above as percentages of the total
df["room_type"].value_counts(normalize=True) * 100

# See unique values in neighbourhood column
df['neighbourhood'].unique()

# Count number of each value in neighbourhood column
df["neighbourhood"].value_counts()

#Crear nuevo dataset con menos columnas

# Creating the new DataFrame with the required columns
listings_prepared_bcn = df[['host_id', 'host_name', 'calculated_host_listings_count', 
                            'host_is_superhost', 'listing_url', 'id', 'name', 
                            'neighbourhood_group_cleansed', 'latitude', 'longitude', 
                            'room_type', 'bedrooms', 'price', 'number_of_reviews', 
                            'review_scores_rating', 'license']].copy()

# Adding 'kitchen', 'patio or balcony', 'elevator', 'air conditioning' columns based on 'amenities'
listings_prepared_bcn['kitchen'] = df['amenities'].str.contains('kitchen', case=False, na=False).astype(int)
listings_prepared_bcn['patio or balcony'] = df['amenities'].str.contains('patio|balcony', case=False, na=False).astype(int)
listings_prepared_bcn['elevator'] = df['amenities'].str.contains('elevator', case=False, na=False).astype(int)
listings_prepared_bcn['air conditioning'] = df['amenities'].str.contains('air conditioning', case=False, na=False).astype(int)

# Adding 'long_term', 'short_term', 'possible_long_term' columns based on 'minimum_nights' and 'maximum_nights'
listings_prepared_bcn['long_term'] = (df['minimum_nights'] > 31).astype(int)
listings_prepared_bcn['short_term'] = (df['maximum_nights'] <= 31).astype(int)
listings_prepared_bcn['possible_long_term'] = ((df['minimum_nights'] >= 1) & (df['maximum_nights'] >= 31)).astype(int)

# Display the first few rows of the new dataframe
print(listings_prepared_bcn.head())


# Export the DataFrame to an Excel file
listings_prepared_bcn.to_excel('listings_prepared_bcn.xlsx', index=False)

# Reordering the columns
ordered_columns = [
    'id', 'listing_url', 'name', 'host_id', 'host_name', 'calculated_host_listings_count', 
    'host_is_superhost', 'license', 'neighbourhood_group_cleansed', 'latitude', 
    'longitude', 'room_type', 'bedrooms', 'price', 'kitchen', 'patio or balcony', 
    'elevator', 'air conditioning', 'long_term', 'short_term', 'possible_long_term', 
    'number_of_reviews', 'review_scores_rating'
]

# Apply the new order to the DataFrame
listings_prepared_bcn = listings_prepared_mad[ordered_columns]

# Display the first few rows of the reordered DataFrame
print(listings_prepared_bcn.head())

# Export the reordered DataFrame to a CSV file
listings_prepared_bcn.to_csv('listings_prepared_bcn.csv', index=False)
