==> 1_Data_Cleaning_BCN.ipynb <==
from pyspark.sql import SparkSession
import pandas as pd

# Initialize a Spark session
spark = SparkSession.builder.appName("AirbnbPriceFilling").getOrCreate()

# Upload Bcn dataset, the original one and the selected listing
orig_bcn = spark.read.csv('Datasets/Dataset_Originales/listings-detailed-bcn-original.csv', header=True, inferSchema=True, multiLine=True, sep=',', quote='"', escape='\\')
df_bcn = spark.read.csv('Datasets/Data_Cleaning_Temporary/listings_prepared_bcn.csv', header=True, inferSchema=True)

df_bcn.show(5)

orig_bcn.printSchema()
from pyspark.sql.functions import col, regexp_extract

# Regular expression to match only numeric strings
df_numeric_ids = orig_bcn.filter(regexp_extract(col("id"), r'^\d+$', 0) != '')

# Show the result to verify the changes
df_numeric_ids.show(5)

# Save the cleaxned DataFrame to a new CSV file
df_numeric_ids.write.csv("temp_ids_bcn.csv", header=True)

df_numeric_ids.count()

# Display the schema of the DataFrame
df_bcn.printSchema()

# Get the number of rows and columns
print(f"Number of rows: {df_bcn.count()}, Number of columns: {len(df_bcn.columns)}")

# Remove the dollar sign and commas, and convert the column to a numerical format:
from pyspark.sql.functions import regexp_replace, col

df_bcn = df_bcn.withColumn("price", regexp_replace(col("price"), "[\$,]", "").cast("float"))

#  Identify Missing Values
from pyspark.sql.functions import col, sum as _sum

# Count missing values for each column
df_bcn.select([_sum(col(c).isNull().cast("int")).alias(c) for c in df_bcn.columns]).show()

### **The dataset has several columns with missing values:**

host_name: Missing in 2 entries > fill in with 'No Disponible'.

host_is_superhost: Missing in 124 entries. > Fill in with 'f'.

license: Missing in 6137 entries (almost all of them). Drop all columns.

neighbourhood_group_cleansed: 1 missing > manual review.

bedrooms: Missing in 1882 entries. > check for more than one bedroom > mark 1, if less than 1 > mark 0

price: Missing in 3736 entries. Drop or fill in depending on the cases.

review_scores_rating: Missing in 5,856 entries. > fill in with 0

# Fill host_name with 'No Disponible'
df_bcn = df_bcn.na.fill({"host_name": "No Disponible"})

# Fill host_is_superhost with 'f'
df_bcn = df_bcn.na.fill({"host_is_superhost": "f"})

# Drop the license Column
df_bcn = df_bcn.drop("license")

# Check by the URL link to fill in the neighbourhood for missing value

# Filter the DataFrame to find the row with the missing value in 'neighbourhood_group_cleansed'
missing_neighbourhood = df_bcn.filter(col("neighbourhood_group_cleansed").isNull())

# Show the row(s) with the missing value
missing_neighbourhood.show()

# Drop the id = 27217399 as it only provides long term rental and it not available for the next 6 months
df_bcn = df_bcn.na.drop(subset=['neighbourhood_group_cleansed'])

### To fill in "bedrooms" columns and also encode 'bedroom' and 'room type' with the following thoery

- If the type is "shared bedroom" or null, we encode it as 0
- If the typr is "private room" or "hotel room", we encode it as 1
- If the type if "entire home/apt", we encode it with 2

from pyspark.sql.functions import when, col

# Apply the Encoding Based on room_type
# Apply the encoding to the 'room_type' column
df_bcn = df_bcn.withColumn("room_type_encoded", 
                   when(col("room_type").isNull(), 0)
                   .when(col("room_type") == "Shared room", 0)
                   .when(col("room_type") == "Private room", 1)
                   .when(col("room_type") == "Hotel room", 1)
                   .when(col("room_type") == "Entire home/apt", 2)
                   .otherwise(0))

# Combine with the 'bedrooms' logic
df_bcn = df_bcn.withColumn("bedrooms_encoded", 
                   when(col("bedrooms") > 1, 1).otherwise(0))

df_bcn.select("room_type", "room_type_encoded", "bedrooms", "bedrooms_encoded").show(10)

# Remove 'room_type' and 'bedrooms'
df_bcn = df_bcn.drop("room_type")
df_bcn = df_bcn.drop("bedrooms")


# Fill review_scores_rating with 0

df_bcn =df_bcn.na.fill({"review_scores_rating": 0})

### Fill in "price" column with the following theory

1. Mark the rows where the "availaility_30" and "availableity_60" are both '0' in the orig_bcn dataframe
2. Found the correspondent rows by using its "id" in df_bcn and drop the entrie rows
3. Fill in the rest of the null values with the everage price in the same neighbourhood

# Load the orig_bcn dataframe
zero_availability_ids = df_numeric_ids.filter((col("availability_30") == 0) & (col("availability_60") == 0)).select("id")
zero_availability_ids.count()

# # Convert zero_availability_ids to a list
zero_availability_ids_list = [row.id for row in zero_availability_ids.distinct().collect()]

# # Filter df_bcn to remove rows where id is in zero_availability_ids_list
df_bcn_filtered = df_bcn.filter(~col("id").isin(zero_availability_ids_list))

df_bcn_filtered.count()

from pyspark.sql import functions as F

# # Calculate the average price per neighbourhood
avg_price_neighbourhood = df_bcn_filtered.groupBy("neighbourhood_group_cleansed").agg(F.round(F.mean("price")).alias("avg_price"))

# Join the original DataFrame with the average prices DataFrame
df_bcn_with_avg = df_bcn_filtered.join(avg_price_neighbourhood, on="neighbourhood_group_cleansed", how="left")

# # Fill missing price values with the calculated average price
df_bcn_final = df_bcn_with_avg.withColumn("price", F.coalesce(col("price"), col("avg_price")))

# # Drop the avg_price column as it's no longer needed
df_bcn_final = df_bcn_final.drop("avg_price")

# # Show the result to verify the changes
df_bcn_final.select("id", "neighbourhood_group_cleansed", "price").show(10)

df_bcn_final.count()

# Count missing values for each column
df_bcn_final.select([_sum(col(c).isNull().cast("int")).alias(c) for c in df_bcn_final.columns]).show()

# Save the cleaned final DataFrame to a new CSV file
df_bcn_final.write.csv("cleaned_listing_bcn.csv", header=True)

spark.stop()
